# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Production work reporting API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /weekly-plans:
    get:
      operationId: listWeeklyPlans
      tags: [weeklyPlans]
      summary: List weekly plans for a week
      parameters:
        - name: weekStart
          in: query
          required: true
          schema:
            type: string
            format: date
          description: ISO date of the Monday (YYYY-MM-DD)
      responses:
        "200":
          description: List of weekly plans with product names
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeeklyPlan"
    post:
      operationId: createWeeklyPlan
      tags: [weeklyPlans]
      summary: Create or update a weekly plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWeeklyPlanBody"
      responses:
        "201":
          description: Created or updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyPlan"
  /weekly-plans/{id}:
    delete:
      operationId: deleteWeeklyPlan
      tags: [weeklyPlans]
      summary: Delete a weekly plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /weekly-plans/progress:
    get:
      operationId: getWeeklyProgress
      tags: [weeklyPlans]
      summary: Get weekly progress for all products
      parameters:
        - name: weekStart
          in: query
          required: true
          schema:
            type: string
            format: date
          description: ISO date of the Monday (YYYY-MM-DD)
      responses:
        "200":
          description: Weekly progress per product
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeeklyProgress"
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        revision:
          type: string
          nullable: true
          maxLength: 10
        createdAt:
          type: string
          format: date-time
...
    WeeklyPlan:
      type: object
      required: [id, productId, weekStart, plannedQuantity, createdAt, updatedAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        weekStart:
          type: string
          format: date
        plannedQuantity:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        product:
          $ref: "#/components/schemas/Product"
    CreateWeeklyPlanBody:
      type: object
      required: [productId, weekStart, plannedQuantity]
      properties:
        productId:
          type: integer
        weekStart:
          type: string
          format: date
        plannedQuantity:
          type: integer
    WeeklyProgress:
      type: object
      required: [productId, productName, plannedQuantity, completedQuantity, remainingQuantity, percentageComplete]
      properties:
        productId:
          type: integer
        productName:
          type: string
        plannedQuantity:
          type: integer
        completedQuantity:
          type: integer
        remainingQuantity:
          type: integer
        percentageComplete:
          type: integer
        planId:
          type: integer
          nullable: true